feat(YATF-99): Wave 3 — Account Detail Dialog
- Create AccountDetailDialog (full-screen) with AccountCards, NetWorthChart, AccountBreakdownChart
- Strip account cards and charts from DashboardPage, replace with dialog trigger
- Simplify RecapCards: remove toggle/accountDetails props, use onOpenAccountDialog
